* Fix bug, inventory items will start decay only at login time * Fix some compiler warnings

+++ otserv/trunk/item.cpp
@@ -846,7 +846,7 @@ uint32_t ItemAttributes::getIntAttr(itemAttrTypes type) const
 		
 	Attribute* attr = getAttr(type);
 	if(attr){
-		return (uint32_t)attr->value;
+		return (uint32_t)(long)attr->value;
 	}
 	else{
 		return 0;
@@ -871,7 +871,7 @@ void ItemAttributes::increaseIntAttr(itemAttrTypes type, uint32_t value)
 	
 	Attribute* attr = getAttr(type);
 	if(attr){
-		attr->value = (void*)((uint32_t)attr->value + value);
+		attr->value = (void*)((long)attr->value + value);
 	}
 }
