+++ otserv/trunk/tools.cpp
@@ -33,10 +33,10 @@ bool fileExists(const char* filename)
 
 void replaceString(std::string& str, const std::string sought, const std::string replacement)
 {
-	uint32_t pos = 0;
-	uint32_t start = 0;
-	uint32_t soughtLen = sought.length();
-	uint32_t replaceLen = replacement.length();
+	size_t pos = 0;
+	size_t start = 0;
+	size_t soughtLen = sought.length();
+	size_t replaceLen = replacement.length();
 	while((pos = str.find(sought, start)) != std::string::npos){
 		str = str.substr(0, pos) + replacement + str.substr(pos + soughtLen);
 		start = pos + replaceLen;
